Fixed the empty PDF data error.
This error occurred every time that we used ZFIX = YES
This meant that the redshift was fixed and LePHARE was fitting a model to that particular redshift. Hence there is no Probability Distribution Function for redshift, and therefore python could not create a plot from an empty array. In this fix, I just added an IF condition to line 190. Simple fix. :)